# Refs #28459 -- Improved performance of SQLCompiler.results_iter().